Added an `editPanel` flow function
... and moved the internals of `addPanel` to a common function for use by both
